ddcmp: Add support for search and replace
This feature allows to search and replace a string in the disk image
to be written. An example would be to modify the cmdline of a kernel
on the fly.

The string length of the string to search must match the length of the
string to replace.

WARNING: If the string to be searched crosses page boundaries, it wont
be replaced!

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>